+++ tests/testthat/test_AutoTuner.R
@@ -5,6 +5,7 @@ test_that("AutoTuner / train+predict", {
   ms = MeasureDummyCPClassif$new(fun = function(pv) if (pv$cp == 0.2) 0 else 1) # lets fake a measure, so we control the best config
   tuner = tnr("grid_search", resolution = 3)
   at = AutoTuner$new(lrn("classif.rpart"), rsmp("holdout"), ms, te, tuner = tuner, ps)
+  expect_false("marshal" %in% at$properties)
   expect_learner(at)
   at$train(task)
   expect_learner(at)
@@ -563,6 +564,51 @@ test_that("AutoTuner errors when second test set is not a subset of task ids", {
   )
 
   expect_error(resample(task, at, resampling_outer, store_models = TRUE), "Test set 2")
+})
+
+# Marshal ----------------------------------------------------------------------
+
+test_that("marshalable learner", {
+  task = tsk("iris")
+  at = auto_tuner(
+    tuner = tnr("random_search", batch_size = 2),
+    learner = lrn("classif.debug"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    store_tuning_instance = TRUE
+  )
+  expect_true("marshal" %in% at$properties)
+
+  at$train(task)
+  at$marshal()
+  at$unmarshal()
+  expect_false(at$learner$marshaled)
+
+  expect_learner(at, task = task)
+})
+
+test_that("marshal", {
+  task = tsk("iris")
+  at = auto_tuner(
+    tuner = tnr("random_search", batch_size = 2),
+    learner = lrn("classif.debug"),
+    resampling = rsmp("cv", folds = 3),
+    measure = msr("classif.ce"),
+    term_evals = 4,
+    store_tuning_instance = TRUE
+  )
+
+  at$train(task)
+
+  model = at$model
+  model1 = marshal_model(model)
+  model2 = unmarshal_model(model1)
+  expect_false(model$learner$marshaled)
+  expect_true(model1$marshaled$learner$marshaled)
+  expect_false(model2$learner$marshaled)
+  expect_equal(class(model), class(model2))
+  expect_class(model1, "marshaled")
 })
 
 # Async ------------------------------------------------------------------------