+++ NEWS.md
@@ -12,6 +12,7 @@
 
 * compatibility: Work with new paradox version 1.0.0
 * fix: `TunerIrace` failed with logical parameters and dependencies.
+* Added marshaling support to `AutoTuner`
 
 # mlr3tuning 0.19.2